#23 Started working on the view prize draw page.
