+++ src/app/service/prize-draw.service.ts
@@ -60,4 +60,29 @@ export class PrizeDrawService {
       }),
     );
   }
+
+  getTimeRemaining(targetDate: Date | undefined): string {
+    if (!targetDate) {
+      return '';
+    }
+    const now = new Date();
+    const endsAt = new Date(targetDate);
+    const diff = endsAt.getTime() - now.getTime();
+
+    if (diff < 0) return 'Expired';
+
+    const minutes = Math.floor(diff / 60000);
+    const hours = Math.floor(diff / 3600000);
+    const days = Math.floor(hours / 24);
+
+    if (hours < 1) {
+      return `${minutes} minutes`;
+    }
+    return (
+      `${days} ` +
+      (days === 1 ? 'day' : 'days') +
+      ` ${hours % 24} ` +
+      ((hours % 24) === 1 ? 'hour' : 'hours')
+    );
+  }
 }
